cognitive fixes and indexer module struct fix
